requirements.txt
fastapi==0.80.0
uvicorn
opencv-python
numpy
mongoengine==0.27.0
PyJWT
Pillow==9.2.0
pilgram==1.2.1
pymongo
pytz
tensorflow==2.8.1
pydantic==1.10.2
typing
tf_slim==1.1.0
boto3
